Set the processor frequency to 456 MHz.
By changing the values of the prediv, postdiv, and multiplier, we can change the frequency.

In the Linux kernel `da850.c` file, you can view the divisor and multiplier values for the supported frequencies (do not forget to subtract 1 from those values).

Before this commit, the frequency was 300 MHz.

It was not possible to launch the cpufreq module to change the frequencies in runtime, perhaps the 4.14 kernel is still raw for this, or you need to add the code yourself.