Set the processor frequency to 456 MHz. (#7)
By changing the values of the prediv, postdiv, and multiplier, the frequency is set to 456 MHz

In the Linux kernel `da850.c` file, one can check the divisor and multiplier values for the supported frequencies (do not forget to subtract 1 from those values).

Before this commit, the frequency was 300 MHz.

It is not possible to use the cpufreq module to set the frequency in runtime, perhaps the in other new kernel version it will become possible.